+++ backend/apps/user/tasks/__client__.py
@@ -1,0 +1,4 @@
+from http import client
+import httpx
+
+client = httpx.Client(http2=True)
