+++ backend/apps/user/tasks/__init__.py
@@ -1,1 +1,1 @@
-from .refresh_jwt import *
+from .jwt import *
