+++ backend/apps/user/tasks/__logger__.py
@@ -1,0 +1,3 @@
+import logging
+
+logger = logging.getLogger("huey")
