+++ backend/apps/user/tasks/jwt/__init__.py
@@ -1,0 +1,2 @@
+from .kitsu import refresh_kitsu_jwt
+from .myanimelist import refresh_mal_jwt
